collision detection
added collision detection. If an object is about to move into another object it will stop now. Also changed the move function to return a rect of where is is going to move to. If that move will be successful, then we set the rect of the object to that new rect.